+++ animation_nodes/nodes/falloff/object_controller_falloff.pyx
@@ -7,6 +7,7 @@ from ... base_types import AnimationNode, VectorizedSocket
 
 from . mix_falloffs import MixFalloffs
 from . invert_falloff import InvertFalloff
+from . radial_falloff import RadialFalloff
 from . constant_falloff import ConstantFalloff
 from . interpolate_falloff import InterpolateFalloff
 from . directional_falloff import UniDirectionalFalloff
@@ -14,12 +15,13 @@ from . point_distance_falloff import PointDistanceFalloff
 
 falloffTypeItems = [
     ("SPHERE", "Sphere", "", "NONE", 0),
-    ("DIRECTIONAL", "Directional", "", "NONE", 1)
+    ("DIRECTIONAL", "Directional", "", "NONE", 1),
+    ("RADIAL", "Radial", "", "NONE", 2),
 ]
 
 mixListTypeItems = [
     ("MAX", "Max", "", "NONE", 0),
-    ("ADD", "Add", "", "NONE", 1)
+    ("ADD", "Add", "", "NONE", 1),
 ]
 
 axisDirectionItems = [(axis, axis, "") for axis in ("X", "Y", "Z", "-X", "-Y", "-Z")]
@@ -51,6 +53,10 @@ class ObjectControllerFalloffNode(bpy.types.Node, AnimationNode):
         if self.falloffType == "SPHERE":
             self.newInput("Float", "Offset", "offset", value = 0)
             self.newInput("Float", "Falloff Width", "falloffWidth", value = 1.0)
+
+        if self.falloffType == "RADIAL":
+            self.newInput("Float", "Phase", "phase")
+
         self.newInput("Interpolation", "Interpolation", "interpolation", defaultDrawType = "PROPERTY_ONLY")
         self.newInput("Boolean", "Invert", "invert", value = False)
         self.newOutput("Falloff", "Falloff", "falloff")
@@ -58,10 +64,13 @@ class ObjectControllerFalloffNode(bpy.types.Node, AnimationNode):
     def draw(self, layout):
         col = layout.column()
         col.prop(self, "falloffType", text = "")
-        if self.falloffType == "DIRECTIONAL":
+        if self.falloffType in ("DIRECTIONAL", "RADIAL"):
             col.row().prop(self, "axisDirection", expand = True)
         if self.useObjectList:
             col.prop(self, "mixListType", text = "")
+
+    def drawAdvanced(self, layout):
+        self.invokeFunction(layout, "createAutoExecutionTrigger", text = "Create Execution Trigger")
 
     def getExecutionFunctionName(self):
         if self.useObjectList:
@@ -69,11 +78,15 @@ class ObjectControllerFalloffNode(bpy.types.Node, AnimationNode):
                 return "execute_Sphere_List"
             elif self.falloffType == "DIRECTIONAL":
                 return "execute_Directional_List"
+            elif self.falloffType == "RADIAL":
+                return "execute_Radial_List"
         else:
             if self.falloffType == "SPHERE":
                 return "execute_Sphere"
             elif self.falloffType == "DIRECTIONAL":
                 return "execute_Directional"
+            elif self.falloffType == "RADIAL":
+                return "execute_Radial"
 
     def execute_Sphere_List(self, objects, offset, falloffWidth, interpolation, invert):
         falloffs = []
@@ -97,7 +110,7 @@ class ObjectControllerFalloffNode(bpy.types.Node, AnimationNode):
         matrix = evaluatedObject.matrix_world
         center = matrix.to_translation()
         size = abs(matrix.to_scale().x) + offset
-        return PointDistanceFalloff(center, size-1, falloffWidth)
+        return PointDistanceFalloff(center, size - 1, falloffWidth)
 
     def execute_Directional_List(self, objects, interpolation, invert):
         falloffs = []
@@ -125,6 +138,30 @@ class ObjectControllerFalloffNode(bpy.types.Node, AnimationNode):
 
         return UniDirectionalFalloff(location, direction, size)
 
+    def execute_Radial_List(self, objects, phase, interpolation, invert):
+        falloffs = []
+        for object in objects:
+            if object is not None:
+                falloffs.append(self.getRadialFalloff(object, phase))
+        if len(falloffs) == 0:
+            return ConstantFalloff(0)
+        else:
+            return self.applyMixInterpolationAndInvert(falloffs, interpolation, invert)
+
+    def execute_Radial(self, object, phase, interpolation, invert):
+        falloff = self.getRadialFalloff(object, phase)
+        return self.applyInterpolationAndInvert(falloff, interpolation, invert)
+
+    def getRadialFalloff(self, object, phase):
+        if object is None:
+            return ConstantFalloff(0)
+        
+        evaluatedObject = getEvaluatedID(object)
+        matrix = evaluatedObject.matrix_world
+        origin = matrix.to_translation()
+        normal = eulerToDirection(matrix.to_euler(), self.axisDirection)
+        return RadialFalloff(origin, normal, phase)
+
     def applyMixInterpolationAndInvert(self, falloffs, interpolation, invert):
         if self.mixListType == "MAX":
             falloff = MixFalloffs(falloffs, "MAX")
@@ -139,3 +176,20 @@ class ObjectControllerFalloffNode(bpy.types.Node, AnimationNode):
         falloff = InterpolateFalloff(falloff, interpolation)
         if invert: falloff = InvertFalloff(falloff)
         return falloff
+
+    def createAutoExecutionTrigger(self):
+        if self.useObjectList or self.inputs["Object"].object is None:
+            return
+
+        customTriggers = self.nodeTree.autoExecution.customTriggers
+
+        attrs = []
+        if self.falloffType in ("DIRECTIONAL", "SPHERE"): attrs.append("scale")
+        if self.falloffType in ("DIRECTIONAL", "RADIAL"): attrs.append("rotation_euler")
+        attrs.append("location")
+        
+        if attrs:
+            item = self.nodeTree.autoExecution.customTriggers.new("MONITOR_PROPERTY")
+            item.idType = "OBJECT"
+            item.dataPaths = ",".join(attrs)
+            item.object = self.inputs["Object"].object