+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - Added edges mode in *Mesh Points Scatter* node.
 - Added *Vertex Color Fill* attribute to GP strokes.
 - Added *Radial Falloff* node.
+- Added *Radial* mode to *Object Controller Falloff* node.
+- Added *Create Trigger* operator to *Object Controller Falloff* node.
+- Added *Extract Matrix Basis* node.
 - Added *Normals* output in *Mesh Points Scatter* node.
 
 ### Fixed